feat: HTTP API, job queue, CLI commands and continuous integration
Adds the FastAPI service with job submission, status, artefact download,
health and readiness probes and optional API-key protection, an in-process
job queue, the serve and join CLI commands, and a CI pipeline that lints,
type-checks, tests on two Python versions, validates the Helm chart across
five deployment postures, builds the images, enforces the comment-free
convention and proves transcription completes with every socket sealed.

Wires the speaker count through the CLI and API, removes settings that were
declared but never read, and fixes the model paths reported by the
deployment validation.

+++ src/hansard/domain/audio.py
@@ -35,8 +35,8 @@ class AudioClip:
 
     def extract(self, span: TimeSpan) -> AudioClip:
         local = TimeSpan(span.start - self.offset, span.end - self.offset).clamped(0.0, self.duration)
-        first = int(round(local.start * self.sample_rate))
-        last = int(round(local.end * self.sample_rate))
+        first = round(local.start * self.sample_rate)
+        last = round(local.end * self.sample_rate)
         return replace(self, samples=self.samples[first:last], offset=self.offset + local.start)
 
     def with_samples(self, samples: np.ndarray) -> AudioClip:
